Revert "refactor(endpoints): rename types.py to models.py to avoid shadowing …"
This reverts commit e2bcfd9efce4a9aea36262ff58ae394af377f707.
