+++ src/mxcp/endpoints/loader.py
@@ -9,7 +9,7 @@ import yaml
 from jsonschema import validate
 
 from mxcp.config.site_config import SiteConfig
-from mxcp.endpoints.models import EndpointDefinition
+from mxcp.endpoints.types import EndpointDefinition
 
 # Configure logging
 logger = logging.getLogger(__name__)
