+++ src/mxcp/endpoints/types.py
